backend/requirements.txt
python-dotenv==1.0.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-Admin==1.6.1
SQLAlchemy==2.0.35
alembic==1.13.3
pyOpenSSL==24.0.0  
psycopg2-binary==2.9.9
pandas==2.2.3
protobuf==5.28.2
Werkzeug==3.0.4
requests==2.32.3
yfinance==0.2.44
authlib==1.3.2
groq==0.11.0